requirements.txt
pillow
requests
pyyaml